backend/djangoapp/requirements.txt
django>=5.0.3,<5.1
djangorestframework>=3.15.1,<3.16
psycopg2-binary>=2.9.9,<2.10
djangorestframework-simplejwt>=5.3.1,<5.4
django-cors-headers>=4.3.1,<4.4